+++ headers/EletricField.hpp
@@ -4,11 +4,11 @@
 class EletricField{
 public:
     EletricField();
-    EletricField(int x,int y,int z);
+    EletricField(double x,double y,double z);
 
     void addNewForceVector(Coord vect);
     void setVectorField(double x, double y, double z);
-    void setPosition(int x,int y,int z);
+    void setPosition(double x,double y,double z);
     
     Coord vectorField;
     Coord position;
